refactored schema destructuring into own component
